Split building.py into a package of four submodules (pure moves)
Phase 2 of the building cleanup (system_docs/building_cleanup_spec.md):
relocate the five classes of the 3,082-line hisim/components/building.py
into hisim/components/building/ by responsibility -- config.py
(BuildingConfig), information.py (BuildingInformation), window.py
(Window), building.py (Building plus its ~30-line BuildingState) -- with
a re-exporting __init__ so the public import path
hisim.components.building.* stays alive for both Python imports and the
fully-qualified classname strings in committed scenario JSONs. Class
bodies are moved verbatim (verified byte-identical against the old
module); the only new lines are module docstrings, '# clean' tags, and
per-module imports.

One backward edge existed: BuildingConfig.get_main_classname returns
Building.get_full_classname(). It is resolved with a runtime-local
import inside that method (module import order stays acyclic), plus the
pylint disables the prospector and critical-only gates require for it.

Scanners checked for flat-module assumptions: check_config_attrs.py
(rglob), json_executor.import_from_string (importlib + getattr) and
project_code_overview_generator.py (os.walk) all handle the package
layout unchanged -- no fixes needed. The phase-1 harness passes with
untouched goldens (they contain no module paths); base suite at its
baseline (3988 passed, 1 skipped); mypy, prospector and the config-attrs
gate are clean.
