+++ hisim/components/building/window.py
@@ -1,0 +1,150 @@
+"""Window solar-gain model used by the Building component.
+
+Part of the ``hisim.components.building`` package split (see the package ``__init__``
+for the layout and the RC_BuildingSimulator reference). Holds ``Window``, moved
+verbatim from the former single-module ``building.py``.
+"""
+
+# clean
+
+import math
+from functools import lru_cache
+
+import pvlib
+
+from hisim import log
+
+
+# =====================================================================================================================================
+class Window:
+    """Based on the RC_BuildingSimulator project @[rc_buildingsimulator-jayathissa] (** Check header)."""
+
+    def __init__(
+        self,
+        window_azimuth_angle=None,
+        window_tilt_angle=None,
+        area=None,
+        glass_solar_transmittance=None,
+        frame_area_fraction_reduction_factor=None,
+        external_shading_vertical_reduction_factor=None,
+        nonperpendicular_reduction_factor=None,
+    ):
+        """Construct all the neccessary attributes."""
+        self.warning_message_already_shown = False
+        # Angles
+        self.window_tilt_angle = window_tilt_angle
+        self.window_azimuth_angle = window_azimuth_angle
+        self.window_tilt_angle_rad: float = 0
+
+        # Area
+        self.area = area
+
+        # Transmittance
+        self.glass_solar_transmittance = glass_solar_transmittance
+        # Incident Solar Radiation
+        self.incident_solar: int
+
+        # Reduction factors
+        self.nonperpendicular_reduction_factor = nonperpendicular_reduction_factor
+        self.external_shading_vertical_reduction_factor = external_shading_vertical_reduction_factor
+        self.frame_area_fraction_reduction_factor = frame_area_fraction_reduction_factor
+
+        self.reduction_factor = (
+            glass_solar_transmittance
+            * nonperpendicular_reduction_factor
+            * external_shading_vertical_reduction_factor
+            * (1 - frame_area_fraction_reduction_factor)
+        )
+
+        self.reduction_factor_with_area = self.reduction_factor * self.area
+
+    def calc_direct_solar_factor(
+        self,
+        sun_altitude,
+        sun_azimuth,
+        apparent_zenith,
+    ):
+        """Calculate the cosine of the angle of incidence on the window.
+
+        Commented equations, that provide a direct calculation, were derived in:
+
+        Proportion of the radiation incident on the window (cos of the incident ray)
+        ref:Quaschning, Volker, and Rolf Hanitsch. "Shade calculations in photovoltaic systems."
+        ISES Solar World Conference, Harare. 1995.
+
+        Based on the RC_BuildingSimulator project @[rc_buildingsimulator-jayathissa] (** Check header)
+        """
+        sun_altitude_rad = math.radians(sun_altitude)
+
+        aoi = pvlib.irradiance.aoi(
+            self.window_tilt_angle,
+            self.window_azimuth_angle,
+            apparent_zenith,
+            sun_azimuth,
+        )
+
+        direct_factor = math.cos(aoi) / (math.sin(sun_altitude_rad))
+
+        return direct_factor
+
+    def calc_diffuse_solar_factor(
+        self,
+    ):
+        """Calculate the proportion of diffuse radiation.
+
+        Based on the RC_BuildingSimulator project @[rc_buildingsimulator-jayathissa] (** Check header)
+        """
+        self.window_tilt_angle_rad = math.radians(self.window_tilt_angle)
+        # Proportion of incident light on the window surface
+        return (1 + math.cos(self.window_tilt_angle_rad)) / 2
+
+    # Calculate solar heat gain through windows.
+    # (** Check header)
+    @lru_cache(maxsize=16)
+    def calc_solar_heat_gains(
+        self,
+        sun_azimuth,
+        direct_normal_irradiance,
+        direct_horizontal_irradiance,
+        global_horizontal_irradiance,
+        direct_normal_irradiance_extra,
+        apparent_zenith,
+        window_tilt_angle,
+        window_azimuth_angle,
+        reduction_factor_with_area,
+    ):
+        """Calculate the Solar Gains in the building zone through the set Window.
+
+        :param sun_altitude: Altitude Angle of the Sun in Degrees
+        :type sun_altitude: float
+        :param sun_azimuth: Azimuth angle of the sun in degrees
+        :type sun_azimuth: float
+        :param normal_direct_radiation: Normal Direct Radiation from weather file
+        :type normal_direct_radiation: float
+        :param horizontal_diffuse_radiation: Horizontal Diffuse Radiation from weather file
+        :type horizontal_diffuse_radiation: float
+        :return: self.incident_solar, Incident Solar Radiation on window
+        :return: self.solar_gains - Solar gains in building after transmitting through the window
+        :rtype: float
+        """
+        if window_azimuth_angle is None:
+            window_azimuth_angle = 0
+            if self.warning_message_already_shown is False:
+                log.warning("window azimuth angle was set to 0 south because no value was set.")
+                self.warning_message_already_shown = True
+
+        poa_irrad = pvlib.irradiance.get_total_irradiance(
+            window_tilt_angle,
+            window_azimuth_angle,
+            apparent_zenith,
+            sun_azimuth,
+            direct_normal_irradiance,
+            global_horizontal_irradiance,
+            direct_horizontal_irradiance,
+            direct_normal_irradiance_extra,
+        )
+
+        if math.isnan(poa_irrad["poa_direct"]):
+            return 0
+
+        return poa_irrad["poa_direct"] * reduction_factor_with_area
